+++ frontend/src/component/incomingWebhooks/IncomingWebhooksModal/IncomingWebhooksModal.tsx
@@ -1,4 +1,4 @@
-import { FormEvent, useEffect, useState } from 'react';
+import { FormEvent, useEffect } from 'react';
 import { Button, styled } from '@mui/material';
 import { SidebarModal } from 'component/common/SidebarModal/SidebarModal';
 import useUiConfig from 'hooks/api/getters/useUiConfig/useUiConfig';
@@ -12,12 +12,11 @@ import {
     useIncomingWebhooksApi,
 } from 'hooks/api/actions/useIncomingWebhooksApi/useIncomingWebhooksApi';
 import { useIncomingWebhookTokensApi } from 'hooks/api/actions/useIncomingWebhookTokensApi/useIncomingWebhookTokensApi';
+import { IncomingWebhooksForm } from './IncomingWebhooksForm/IncomingWebhooksForm';
 import {
-    IncomingWebhooksFormErrors,
-    IncomingWebhooksForm,
     TokenGeneration,
-    DEFAULT_INCOMING_WEBHOOKS_FORM_ERRORS,
-} from './IncomingWebhooksForm/IncomingWebhooksForm';
+    useIncomingWebhooksForm,
+} from './IncomingWebhooksForm/useIncomingWebhooksForm';
 
 const StyledForm = styled('form')(() => ({
     display: 'flex',
@@ -49,76 +48,48 @@ export const IncomingWebhooksModal = ({
     setOpen,
     newToken,
 }: IIncomingWebhooksModalProps) => {
-    const { incomingWebhooks, refetch } = useIncomingWebhooks();
+    const { refetch } = useIncomingWebhooks();
     const { addIncomingWebhook, updateIncomingWebhook, loading } =
         useIncomingWebhooksApi();
     const { addIncomingWebhookToken } = useIncomingWebhookTokensApi();
     const { setToastData, setToastApiError } = useToast();
     const { uiConfig } = useUiConfig();
 
-    const [enabled, setEnabled] = useState(false);
-    const [name, setName] = useState('');
-    const [description, setDescription] = useState('');
-    const [tokenGeneration, setTokenGeneration] = useState<TokenGeneration>(
-        TokenGeneration.LATER,
-    );
-    const [tokenName, setTokenName] = useState('');
-    const [errors, setErrors] = useState<IncomingWebhooksFormErrors>(
-        DEFAULT_INCOMING_WEBHOOKS_FORM_ERRORS,
-    );
-    const [validated, setValidated] = useState(false);
+    // const {
+    //     enabled,
+    //     setEnabled,
+    //     name,
+    //     setName,
+    //     description,
+    //     setDescription,
+    //     tokenGeneration,
+    //     setTokenGeneration,
+    //     tokenName,
+    //     setTokenName,
+    //     errors,
+    //     validateName,
+    //     validateTokenName,
+    //     validate,
+    //     validated,
+    //     reloadForm,
+    // } = useIncomingWebhooksForm(incomingWebhook);
+
+    const {
+        enabled,
+        name,
+        description,
+        tokenGeneration,
+        tokenName,
+        validate,
+        reloadForm,
+    } = useIncomingWebhooksForm(incomingWebhook);
 
     useEffect(() => {
-        setEnabled(incomingWebhook?.enabled || false);
-        setName(incomingWebhook?.name || '');
-        setDescription(incomingWebhook?.description || '');
-        setTokenGeneration(TokenGeneration.LATER);
-        setTokenName('');
-        setErrors(DEFAULT_INCOMING_WEBHOOKS_FORM_ERRORS);
-    }, [open, incomingWebhook]);
+        reloadForm();
+    }, [open]);
 
     const editing = incomingWebhook !== undefined;
     const title = `${editing ? 'Edit' : 'New'} incoming webhook`;
-
-    // TODO: useForm / useFormValidation hook?
-    const isEmpty = (value: string) => !value.length;
-    const isNameNotUnique = (value: string) =>
-        !incomingWebhooks?.some(
-            ({ id, name }) => id !== incomingWebhook?.id && name === value,
-        );
-    const validateName = (name: string) => {
-        if (isEmpty(name)) {
-            setError(ErrorField.NAME, 'Name is required.');
-            return false;
-        }
-
-        if (isNameNotUnique(name)) {
-            setError(ErrorField.NAME, 'Name must be unique.');
-            return false;
-        }
-
-        clearError(ErrorField.NAME);
-        return true;
-    };
-    const validateTokenName = (tokenName: string) => {
-        if (tokenGeneration === TokenGeneration.NOW && isEmpty(tokenName)) {
-            setError(ErrorField.TOKEN_NAME, 'Token name is required.');
-            return false;
-        }
-
-        clearError(ErrorField.TOKEN_NAME);
-        return true;
-    };
-    const validate = () => {
-        const validName = validateName(name);
-        const validTokenName = validateTokenName(tokenName);
-
-        setValidated(true);
-
-        return validName && validTokenName;
-    };
-
-    const showErrors = validated && Object.values(errors).some(Boolean);
 
     const payload: IncomingWebhookPayload = {
         enabled,
@@ -185,20 +156,21 @@ export const IncomingWebhooksModal = ({
                 <StyledForm onSubmit={onSubmit}>
                     <IncomingWebhooksForm
                         incomingWebhook={incomingWebhook}
-                        enabled={enabled}
-                        name={name}
-                        description={description}
-                        tokenGeneration={tokenGeneration}
-                        tokenName={tokenName}
-                        setEnabled={setEnabled}
-                        setName={setName}
-                        setDescription={setDescription}
-                        setTokenGeneration={setTokenGeneration}
-                        setTokenName={setTokenName}
-                        errors={errors}
-                        showErrors={showErrors}
-                        validateName={validateName}
-                        validateTokenName={validateTokenName}
+                        // incomingWebhook={incomingWebhook}
+                        // enabled={enabled}
+                        // name={name}
+                        // description={description}
+                        // tokenGeneration={tokenGeneration}
+                        // tokenName={tokenName}
+                        // setEnabled={setEnabled}
+                        // setName={setName}
+                        // setDescription={setDescription}
+                        // setTokenGeneration={setTokenGeneration}
+                        // setTokenName={setTokenName}
+                        // errors={errors}
+                        // showErrors={showErrors}
+                        // validateName={validateName}
+                        // validateTokenName={validateTokenName}
                     />
                     <StyledButtonContainer>
                         <Button
